+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## Added
+
+- 1080p video conversion preset
+
+## Changed
+
+- Videos with height >= 1080 will now re-encoded also with the 1080p preset
+- The original video is not passed anymore to the shaka packager
+
+
 ## [0.2.0] - 2017-09-27
 
 ## Added